fixed parsing /proc/$pid/stat starttime field
On systems that have been running for a long time (for example 552
days) we were failing parsing the starttime field:

```
Could not find or convert Starttime. This should never happen.
Please report this incident to the Opensnitch developers:
strconv.Atoi: parsing "4242026842": value out of range
```

- extra: fixed tests.

+++ daemon/procmon/parse.go
@@ -92,7 +92,7 @@ func FindProcess(pid int, interceptUnknown bool) *Process {
 		return NewProcess(0, "")
 	}
 
-	if proc := findProcessInActivePidsCache(uint32(pid)); proc != nil {
+	if proc := findProcessInActivePidsCache(uint64(pid)); proc != nil {
 		return proc
 	}
 
@@ -110,7 +110,7 @@ func FindProcess(pid int, interceptUnknown bool) *Process {
 			proc.readEnv()
 			proc.cleanPath()
 
-			addToActivePidsCache(uint32(pid), proc)
+			addToActivePidsCache(uint64(pid), proc)
 			return proc
 		}
 	}
@@ -128,7 +128,7 @@ func FindProcess(pid int, interceptUnknown bool) *Process {
 		proc.readEnv()
 		proc.cleanPath()
 
-		addToActivePidsCache(uint32(pid), proc)
+		addToActivePidsCache(uint64(pid), proc)
 		return proc
 	}
 	return nil
